% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgsflaGrassStat.R
\docType{package}
\name{getArgsflaGrassStat}
\alias{getArgsflaGrassStat}
\alias{getArgsflaGrassStat-package}
\title{getArgsflaGrassStat package}
\description{
Returning a dataframe with arguments after reading from an ini-file or from command prompt. Part of project calculating statistics from shape-points and interpreted satelite-images using Grass and GDAL, Today called flaGrassStat.
}
\section{getArgsflaGrassStat functions}{

read_ini_file
read_arguments
}

\author{
Folke Larsson in Boden Sweden
}
