% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgsGstat.R
\docType{package}
\name{getArgsGstat}
\alias{getArgsGstat}
\alias{getArgsGstat-package}
\title{getArgsGstat package}
\description{
Returning a dataframe with arguments after reading from an ini-file or from command prompt. Part of project calculating statistics from shape-points and interpreted satelite-images using Grass and GDAL, Today called flaGrassStat.
}
\section{getArgdGstat functions}{

read_ini_file
read_arguments
}

\author{
Folke Larsson \, Boden Sweden
}
