% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArguments.R
\name{read_arguments}
\alias{read_arguments}
\title{read_arguments ini-file command prompt}
\usage{
read_arguments(v_filepath, v_filename, v_read_from_inifile, v_task)
}
\arguments{
\item{v_filepath}{path to file}

\item{v_filename}{name of the file}

\item{v_read_from_inifile}{using inifile(TRUE) command line(FALSE)}

\item{v_task}{is the task parameters are supporting}
}
\value{
dataframe with processed arguments
}
\description{
reads parameters from an ini-file or command prompt into a dataframe that it returns. part of specific project.
}
\examples{
/dontrun{
df_args <- read_arguments("file path", "params.ini", TRUE, "calculate_means")
}
}
\author{
Folke Larsson in Boden Sweden
}
