% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIniFile.R
\name{read_ini_file}
\alias{read_ini_file}
\title{read_ini_file parameters}
\usage{
read_ini_file(v_path, v_filename)
}
\arguments{
\item{v_path}{path to file}

\item{v_filename}{name of the file}
}
\value{
dataframe with rows conaining arguments
}
\description{
read variables from an ini-file into a dataframe that it returns. Part of specific project.
}
\examples{
/donttest {
	df_argument_rows <- read_ini_file("/grass7/parameter_files", "parameters.ini")
}
}
\author{
Folke Larsson Boden Sweden
}
